cleaning up texture structure

+++ src/engine/util/random/random.hpp
@@ -5,9 +5,9 @@ namespace Engine::Random
 {
     void internal_engine_init();
 
-    float rangeI(float min, float max);
-    double rangeF(double min, double max);
-    int rangeD(int min, int max);
+    int rangeI(int min, int max);
+    float rangeF(float min, float max);
+    double rangeD(double min, double max);
 }
 
 #endif // SRC_ENGINE_UTIL_RANDOM_RANDOM
